Modified directory structure of language model
