+++ espnet_onnx/export/asr/models/language_models/embed.py
@@ -8,6 +8,67 @@
 import math
 import torch
 
+from espnet.nets.pytorch_backend.transformer.subsampling import Conv2dSubsampling
+from espnet.nets.pytorch_backend.transformer.subsampling import Conv2dSubsampling2
+from espnet.nets.pytorch_backend.transformer.subsampling import Conv2dSubsampling6
+from espnet.nets.pytorch_backend.transformer.subsampling import Conv2dSubsampling8
+from espnet.nets.pytorch_backend.transformer.embedding import (
+    PositionalEncoding,
+    ScaledPositionalEncoding,
+    RelPositionalEncoding,
+    LegacyRelPositionalEncoding,
+    StreamPositionalEncoding,
+)
+
+
+def get_pos_emb(pos_emb):
+    if isinstance(pos_emb, LegacyRelPositionalEncoding):
+        return OnnxLegacyRelPositionalEncoding(pos_emb)
+    elif isinstance(pos_emb, ScaledPositionalEncoding):
+        return OnnxScaledPositionalEncoding(pos_emb)
+    elif isinstance(pos_emb, RelPositionalEncoding):
+        return OnnxRelPositionalEncoding(pos_emb)
+    elif isinstance(pos_emb, PositionalEncoding):
+        return OnnxPositionalEncoding(pos_emb)
+    elif isinstance(pos_emb, StreamPositionalEncoding):
+        return OnnxStreamPositionalEncoding(pos_emb)
+    elif (isinstance(pos_emb, nn.Sequential) and len(pos_emb) == 0) \
+        or (isinstance(pos_emb, Conv2dSubsamplingWOPosEnc)):
+        return pos_emb
+    else:
+        raise ValueError('Embedding model is not supported.')
+
+
+class Embedding(nn.Module):
+    def __init__(self, model):
+        super().__init__()
+        self.model = model
+
+        if not isinstance(model, nn.Embedding):
+            if (
+                isinstance(model, Conv2dSubsampling)
+                or isinstance(model, Conv2dSubsampling2)
+                or isinstance(model, Conv2dSubsampling6)
+                or isinstance(model, Conv2dSubsampling8)
+            ):
+                self.model.out[-1] = get_pos_emb(model.out[-1])
+            else:
+                self.model[-1] = get_pos_emb(model[-1])
+
+    def forward(self, x, mask=None):
+        if mask is None:
+            return self.model(x)
+        else:
+            xs = self.model(x, mask)
+            if isinstance(self.model, Conv2dSubsampling):
+                return xs, mask[:, :, :-2:2][:, :, :-2:2]
+            elif isinstance(self.model, Conv2dSubsampling2):
+                return xs, mask[:, :, :-2:2][:, :, :-2:1]
+            elif isinstance(self.model, Conv2dSubsampling6):
+                return xs, mask[:, :, :-2:2][:, :, :-4:3]
+            elif isinstance(self.model, Conv2dSubsampling8):
+                return xs, mask[:, :, :-2:2][:, :, :-2:2][:, :, :-2:2]
+
 
 def _pre_hook(
     state_dict,
